apply ts on PerfumeSummaryDTO
